Image Captioning/requirements.txt
tensorflow
numpy
streamlit
pillow
matplotlib
